requirements.txt
streamlit
tensorflow==2.10.0
numpy==1.23.5
pillow